Fix JAVA_HOME used by gradle cmakeBuild task
gradle is changing the JAVA_HOME path to be the nested JRE path instead of the java home set by the builder.
Then the build fails because the JRE path doesn't contain the JNI include headers and libraries.

This commit creates an optional override to the JAVA_HOME used
by the cmakeBuild task.