Adjusted memory trace and dump to use static allocator directly.
Instead of using aws_jni_get_allocator, the static allocator is checked and used
directly. This prevents the possibility of the JNI layer from creating a new
allocator just to check a memory dump or trace.

If the allocator does not exist, then one can safely assume it has freed all memory
prior to the allocator being deleted.
This seems, based on the 12 tests I've run so far, fix the memory leak.