Clean up .dll files on Windows (#493)
**ISSUE:** On Windows, the .dll file extracted to the temp directory was never being deleted. And repeated runs of the program were filling the disk with old .dll files.

**DESCRIPTION OF CHANGES:** On Windows only, during startup, scan the temp dir for old instances of `AWSCRT_*aws-crt-jni.dll` and try to delete them.

**DIAGNOSIS:** We were using [File.deleteOnExit()](https://docs.oracle.com/javase/8/docs/api/java/io/File.html#deleteOnExit--) to clean up the shared lib, which is working OK on other platforms, but not on Windows. On Windows, files cannot be deleted while they're in use, and it appears that Java doesn't try to "unload" the .dll, so it's apparently impossible for a Java process to delete a .dll that it's using.

I did an experiment with a dead-simple JNI library. I simply loaded it, and called `File.deleteOnExit()`. That .dll didn't get deleted either. So it's not just a problem with our .dll being very complex.

**SIDE-NOTE:**
I realized that on non-Windows platforms, the shared-lib won't be deleted on exit if Java dies suddenly (crash in C, OS terminated process, etc). But on non-Windows platforms you can delete the shared-lib from disk immediately after loading it, you don't need to wait until the process exits. I did experiments on Mac and this works. But I left it out of this PR because this is an emergency fix for windows and I don't want to introduce non-necessary changes.