HTTP handles early response with "Connection: close" header (#597)
*Issue #:* https://github.com/awslabs/aws-c-s3/issues/264

*Description of changes:*
Fix issue where an HTTP connection closing after the response was received, but before the request was completely sent, was treated as an error. An HTTP exchange is now considered successful as long as the complete response was received. This fix was made in aws-c-http: https://github.com/awslabs/aws-c-http/pull/431