Do not require `coreLibraryDesugaring` on Android
The `java.time` package is only available natively on Android API 28+.
Using it forces consuming apps to enable `coreLibraryDesugaring` to
support older API levels.

This library currently uses `java.time.Instant` only to log the current
date, which can be done easily using the old `java.util.Date`.

Do note that the exact output of `Instant.toString()` depends on the
JDK version, so the resulting string is not necessarily the same, but
the current one should be OK for logging purposes.